+++ lib/src/rx/rx_value_extension.dart
@@ -1,35 +1,33 @@
 part of view_model_kit;
 
-extension CreateReactiveValueWithViewModel on BaseViewModel {
+abstract interface class BaseRxCreator {}
+
+extension CreateReactiveValueWithRContainer on BaseRxCreator {
   /// create Reactive Value (int, double, string, model(value object), ...)
-  R<V> create<V>(V value) {
-    final rx = R._(value);
-    rx._notifyAtStatefulWidget = notify;
-    _addRxDisposer(rx._dispose);
+  R<V> create<V>(V value) => createMutable<V>(value);
+
+  MutableR<V> createMutable<V>(V value) {
+    final rx = MutableR._(value);
+    if (this is BaseViewModel) {
+      rx._notifyAtStatefulWidget = (this as BaseViewModel).notify;
+      (this as BaseViewModel)._addRxDisposer(rx._dispose);
+    } else {
+      rx._notifyAtStatefulWidget = null;
+    }
     return rx;
   }
 
   /// create Reactive List
-  RList<V> createList<V>([List<V>? list]) {
-    final rx = RList._(list);
-    rx._notifyAtStatefulWidget = notify;
-    _addRxDisposer(rx._dispose);
-    return rx;
-  }
-}
+  RList<V> createList<V>([List<V>? list]) => createMutableList(list);
 
-extension CreateReactiveValueWithRContainer on BaseContainer {
-  /// create Reactive Value (int, double, string, model(value object), ...)
-  R<V> create<V>(V value) {
-    final rx = R._(value);
-    rx._notifyAtStatefulWidget = null;
-    return rx;
-  }
-
-  /// create Reactive List
-  RList<V> createList<V>([List<V>? list]) {
-    final rx = RList._(list);
-    rx._notifyAtStatefulWidget = null;
+  MutableRList<V> createMutableList<V>([List<V>? list]) {
+    final rx = MutableRList._(list);
+    if (this is BaseViewModel) {
+      rx._notifyAtStatefulWidget = (this as BaseViewModel).notify;
+      (this as BaseViewModel)._addRxDisposer(rx._dispose);
+    } else {
+      rx._notifyAtStatefulWidget = null;
+    }
     return rx;
   }
 }