+++ lib/src/rx/rx_value_impl.dart
@@ -1,7 +1,9 @@
 part of view_model_kit;
 
 /// Reactive Value (int, double, string, model(value object), ...)
-class R<V> extends BaseR<V> {
+///
+/// (immutable, read only)
+abstract interface class R<V> extends BaseR<V> {
   V _value;
 
   @override
@@ -12,11 +14,20 @@ class R<V> extends BaseR<V> {
   @override
   V get value => _value;
 
+  @override
+  String toString() => "R($_value)";
+
+  MutableR<V> toMutable() => this as MutableR<V>;
+}
+
+/// Reactive Value (int, double, string, model(value object), ...)
+///
+/// (mutable, read/write)
+final class MutableR<V> extends R<V> {
+  MutableR._(V value) : super._(value);
+
   set value(V newValue) {
     _value = newValue;
     _notify();
   }
-
-  @override
-  String toString() => "R($_value)";
 }